+++ src/math_utils.h
@@ -21,4 +21,4 @@ Vector3D reflect(const Vector3D& incident, const Vector3D& normal);
  * point - intersection point
  * bias - small shift to the intersection point
 */
-Vector3D refract(const Vector3D& incident, Vector3D& normal, double ior, Vector3D &point, double bias);
+Vector3D refract(const Vector3D& incident, Vector3D& normal, float ior, Vector3D &point, float bias);
