+++ src/timer.h
@@ -9,14 +9,14 @@ template <class Fn, class... Args>
 auto timer(Fn fn, Args && ... args)
     -> typename std::enable_if< 
     	!std::is_same< decltype( fn( std::forward<Args>(args) ... )), void >::value,
-    	std::pair<double, decltype(fn(args...))> >::type
+    	std::pair<float, decltype(fn(args...))> >::type
 {
   static_assert(!std::is_void<decltype(fn(args...))>::value,
                 "Call timer_void if return type is void!");
   auto start = std::chrono::high_resolution_clock::now();
   auto ret = fn(args...);
   auto end = std::chrono::high_resolution_clock::now();
-  std::chrono::duration<double> elapsed_seconds = end - start;
+  std::chrono::duration<float> elapsed_seconds = end - start;
   return { elapsed_seconds.count(), ret };
 }
  
@@ -24,13 +24,13 @@ auto timer(Fn fn, Args && ... args)
 template <class Fn, class ... Args>
 auto timer(Fn fn, Args && ... args) -> typename std::enable_if< 
     	std::is_same< decltype( fn( std::forward<Args>(args) ... )), void >::value,
-    	double>::type
+    	float>::type
 {
   static_assert(std::is_void<decltype(fn(args...))>::value,
                 "Call timer for non void return type");
   auto start = std::chrono::high_resolution_clock::now();
   fn(args...);
   auto end = std::chrono::high_resolution_clock::now();
-  std::chrono::duration<double> elapsed_seconds = end - start;
+  std::chrono::duration<float> elapsed_seconds = end - start;
   return elapsed_seconds.count();
 }
